add fitwide cog
