+++ cogs/verify.py
@@ -1,12 +1,7 @@
-import discord
 from discord.ext import commands
-from sqlalchemy.orm.exc import NoResultFound
-import utils
 from config import config, messages
 from features import verification
 from repository import user_repo
-from repository.database import session
-from repository.database.verification import Valid_person, Permit
 
 user_r = user_repo.UserRepository()
 
@@ -29,109 +24,6 @@ class Verify(commands.Cog):
     async def getcode(self, ctx):
         await self.verification.send_code(ctx.message)
 
-    @commands.cooldown(rate=2, per=20.0, type=commands.BucketType.user)
-    @commands.command()
-    async def role_check(self, ctx, p_verified: bool=True,
-                         p_move: bool=True, p_status: bool=True,
-                         p_role: bool=True, p_muni: bool=True):
-        if ctx.author.id != config.admin_id:
-            await ctx.send(
-                    messages.insufficient_rights
-                    .format(user=utils.generate_mention(ctx.author.id)))
-            return
-
-        guild = self.bot.get_guild(config.guild_id)
-        members = guild.members
-
-        verify = discord.utils.get(guild.roles, name="Verify")
-        host = discord.utils.get(guild.roles, name="Host")
-        bot = discord.utils.get(guild.roles, name="Bot")
-        poradce = discord.utils.get(guild.roles, name="Poradce")
-        dropout = discord.utils.get(guild.roles, name="Dropout")
-        muni = discord.utils.get(guild.roles, name="MUNI")
-
-        verified = [member for member in members
-                    if verify in member.roles and
-                    host not in member.roles and 
-                    bot not in member.roles and 
-                    dropout not in member.roles and 
-                    poradce not in member.roles]
-
-        if not p_muni:
-            verified = [member for member in verified
-                        if muni not in member.roles]
-
-        permited = session.query(Permit)
-        permited_ids = [int(person.discord_ID) for person in permited]
-
-        bit0 = discord.utils.get(guild.roles, name="0BIT")
-        bit1 = discord.utils.get(guild.roles, name="1BIT")
-        bit2 = discord.utils.get(guild.roles, name="2BIT")
-        bit3 = discord.utils.get(guild.roles, name="3BIT")
-        bit4 = discord.utils.get(guild.roles, name="4BIT+")
-        mit1 = discord.utils.get(guild.roles, name="1MIT")
-
-        for member in verified:
-            if member.id not in permited_ids:
-                if p_verified:
-                    await ctx.send("Nenasel jsem v verified databazi: " +
-                                   utils.generate_mention(member.id))
-            else:
-                try:
-                    login = session.query(Permit).\
-                        filter(Permit.discord_ID == str(member.id)).one().login
-
-                    person = session.query(Valid_person).\
-                        filter(Valid_person.login == login).one()
-                except NoResultFound:
-                    continue
-
-                if person.status != 0:
-                    if p_status:
-                        await ctx.send("Status nesedi u: " + login)
-
-                year = self.verification.transform_year(person.year)
-
-                role = discord.utils.get(guild.roles, name=year)
-
-                if role not in member.roles:
-                    if year == "1MIT" and bit4 in member.roles and p_move:
-                        await member.add_roles(mit1)
-                        await member.remove_roles(bit4)
-                        await ctx.send("Presouvam: " + member.display_name +
-                                       " do 1MIT")
-                    elif year == "4BIT+" and bit3 in member.roles and p_move:
-                        await member.add_roles(bit4)
-                        await member.remove_roles(bit3)
-                        await ctx.send("Presouvam: " + member.display_name +
-                                       " do 4BIT+")
-                    elif year == "3BIT" and bit2 in member.roles and p_move:
-                        await member.add_roles(bit3)
-                        await member.remove_roles(bit2)
-                        await ctx.send("Presouvam: " + member.display_name +
-                                       " do 3BIT")
-                    elif year == "2BIT" and bit1 in member.roles and p_move:
-                        await member.add_roles(bit2)
-                        await member.remove_roles(bit1)
-                        await ctx.send("Presouvam: " + member.display_name +
-                                       " do 2BIT")
-                    elif year == "1BIT" and bit0 in member.roles and p_move:
-                        await member.add_roles(bit1)
-                        await member.remove_roles(bit0)
-                        await ctx.send("Presouvam: " + member.display_name +
-                                       " do 1BIT")
-                    elif not p_role:
-                        continue
-                    elif year is None:
-                        await ctx.send("Nesedi mi role u: " +
-                                       utils.generate_mention(member.id) +
-                                       ", ma ted rocnik: " + person.year)
-                    else:
-                        await ctx.send("Nesedi mi role u: " + 
-                                       utils.generate_mention(member.id) +
-                                       ", mel by mit roli: " + year)
-
-        await ctx.send("Done")
 
 def setup(bot):
     bot.add_cog(Verify(bot))
